requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
cloudinary==1.44.2
dj-database-url==3.0.1
Django==4.2.30
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.13
packaging==26.1
pillow==11.3.0
psycopg2-binary==2.9.12
PyJWT==2.12.1
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0